requirements.txt
#actual project dependencies
asgiref==3.2.10
Django==3.0.8
install==1.3.3
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.0

## Test and quality analysis
astroid==2.4.2
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1

pep8==1.7.1